FIX: Relocate a header to avoid header test failure.
When compiling LSTK togather with ITK, a header test is automatically
generated for LSTK, which will fail due to itkIncludeRequiredIOFactories.h
being listed in that header test.

LSTK does have different layout than an internal module in ITK. In ITK the src module
usually should not deal with imageIO, only test module need to deal with IO by
registering IO factories through TestKernel module.
The minimum work to do now is to move the header to the src dir.